+++ Source/Bifrost.Specs/Execution/for_AssemblyProvider/when_loading_assembly_again.cs
@@ -12,27 +12,27 @@ namespace Bifrost.Specs.Execution.for_AssemblyProvider
     public class when_loading_assembly_again : given.an_assembly_provider
     {
         static IObservableCollection<Assembly> result;
+        static IObservableCollection<AssemblyInfo> available_assemblies;
         static AssemblyInfo assembly_info;
         static Type[] types;
         static Assembly assembly;
 
         Establish context = () =>
         {
-            assembly_info = new AssemblyInfo("x", "y");
             types = new Type[0];
             assembly = new TestAssembly("x", "location.dll", types);
-            GetMock<ICanProvideAssemblies>().Setup(m => m.AvailableAssemblies).Returns(new[] { assembly_info });
-            GetMock<IAssemblyUtility>().Setup(m => m.IsAssembly(assembly_info)).Returns(true);
-            GetMock<ICanProvideAssemblies>().Setup(m => m.Get(assembly_info)).Returns(assembly);
+            assembly_info = new AssemblyInfo("x", "y", new Lazy<Assembly>(() => assembly));
+            available_assemblies = new ObservableCollection<AssemblyInfo> { assembly_info };
+            GetMock<ICanProvideAssemblies>().Setup(m => m.AvailableAssemblies).Returns(available_assemblies);
+            GetMock<IAssemblySpecifiers>().Setup(m => m.SpecifyUsingSpecifiersFrom(assembly)).Returns(true);
             GetMock<IAssemblyFilters>().Setup(m => m.ShouldInclude("location.dll")).Returns(true);
 
             provider = Get<AssemblyProvider>();
         };
 
-        Because of = () => GetMock<ICanProvideAssemblies>().Raise(x => x.AssemblyAdded += null, assembly);
+        Because of = () => available_assemblies.Add(assembly_info);
 
-        It should_specify_from_the_assembly_once = () =>
-            GetMock<IAssemblySpecifiers>().Verify(m => m.SpecifyUsingSpecifiersFrom(assembly), Times.Once);
+        It should_specify_from_the_assembly = () => GetMock<IAssemblySpecifiers>().VerifyAll();
 
         It should_return_the_assembly_only_once = () => provider.GetAll().ShouldContainOnly(assembly);