+++ Source/Bifrost.Specs/Execution/for_AssemblyProvider/when_providing_from_two_providers_with_same_assembly.cs
@@ -23,17 +23,21 @@ namespace Bifrost.Specs.Execution.for_AssemblyProvider
 
         Establish context = () =>
         {
-            assembly_info1 = new AssemblyInfo("SameName", "y");
-            types1 = new[] {typeof(string)};
             assembly1 = new TestAssembly("x", "location1.dll", types1);
-            assembly_info2 = new AssemblyInfo("SameName", "t");
+            assembly_info1 = new AssemblyInfo("SameName", "y", new Lazy<Assembly>(() => assembly1));
+            types1 = new[] { typeof(string) };
+            assembly_info2 = new AssemblyInfo("SameName", "t", new Lazy<Assembly>(() => null));
             provider1 = new Mock<ICanProvideAssemblies>();
             provider2 = new Mock<ICanProvideAssemblies>();
             RebindMultiple(provider1.Object, provider2.Object);
-            provider1.Setup(m => m.AvailableAssemblies).Returns(new[] {assembly_info1});
-            provider2.Setup(m => m.AvailableAssemblies).Returns(new[] {assembly_info2});
-            GetMock<IAssemblyUtility>().Setup(m => m.IsAssembly(assembly_info1)).Returns(true);
-            provider1.Setup(m => m.Get(assembly_info1)).Returns(assembly1);
+            provider1
+                .Setup(m => m.AvailableAssemblies)
+                .Returns(new ObservableCollection<AssemblyInfo> { assembly_info1 });
+            provider2
+                .Setup(m => m.AvailableAssemblies)
+                .Returns(new ObservableCollection<AssemblyInfo> { assembly_info2 });
+            GetMock<IAssemblySpecifiers>().Setup(m => m.SpecifyUsingSpecifiersFrom(assembly1)).Returns(false);
+            GetMock<IAssemblySpecifiers>().Setup(m => m.SpecifyUsingSpecifiersFrom(assembly1)).Returns(true);
             GetMock<IAssemblyFilters>().Setup(m => m.ShouldInclude("location1.dll")).Returns(true);
             GetMock<IAssemblyFilters>().Setup(m => m.ShouldInclude("location2.dll")).Returns(true);
 
@@ -41,9 +45,6 @@ namespace Bifrost.Specs.Execution.for_AssemblyProvider
         };
 
         Because of = () => result = provider.GetAll();
-
-        It should_not_check_second_assembly = () =>
-            GetMock<IAssemblyUtility>().Verify(m => m.IsAssembly(assembly_info2), Times.Never);
 
         It should_not_filter_second_assembly = () =>
             GetMock<IAssemblyFilters>().Verify(m => m.ShouldInclude("location2.dll"), Times.Never);