+++ Source/Bifrost/Execution/AssemblyProvider.cs
@@ -1,70 +1,69 @@
-﻿#region License
-//
-// Copyright (c) 2008-2015, Dolittle (http://www.dolittle.com)
-//
-// Licensed under the MIT License (http://opensource.org/licenses/MIT)
-//
-// You may not use this file except in compliance with the License.
-// You may obtain a copy of the license at
-//
-//   http://github.com/dolittle/Bifrost/blob/master/MIT-LICENSE.txt
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-#endregion
+﻿using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
+using System.Threading;
 using Bifrost.Collections;
-using Bifrost.Extensions;
 
 namespace Bifrost.Execution
 {
     /// <summary>
-    /// Represents an implementation of <see cref="IAssemblyProvider"/>
+    /// Represents an implementation of <see cref="IAssemblyProvider"/>.
     /// </summary>
     [Singleton]
     public class AssemblyProvider : IAssemblyProvider
     {
-        static readonly object LockObject = new object();
+        int _processing;
 
-        readonly AssemblyComparer _comparer = new AssemblyComparer();
+        // The implementation strategy is to enqueue all discovered assemblies in this queue..
+        readonly ConcurrentQueue<AssemblyInfo> _assembliesToBeSpecified = new ConcurrentQueue<AssemblyInfo>();
 
-        readonly IEnumerable<ICanProvideAssemblies> _assemblyProviders;
+        // ..and after they have been checked for specifiers move them to this list..
+        readonly IList<AssemblyInfo> _specifiedAssemblies = new List<AssemblyInfo>();
+
+        // ..And then, each time a new specification is discovered, all assemblies in the list above
+        // are checked for inclusion and moved to this collection of assemblies if they should be included.
+        readonly ObservableCollection<Assembly> _assemblies = new ObservableCollection<Assembly>();
+
+        readonly AssemblyInfoComparer _comparer = new AssemblyInfoComparer();
+
         readonly IAssemblyFilters _assemblyFilters;
-        readonly IAssemblyUtility _assemblyUtility;
         readonly IAssemblySpecifiers _assemblySpecifiers;
         readonly IContractToImplementorsMap _contractToImplementorsMap;
-        readonly ObservableCollection<Assembly> _assemblies = new ObservableCollection<Assembly>();
 
         /// <summary>
         /// Initializes a new instance of <see cref="AssemblyProvider"/>
         /// </summary>
-        /// <param name="assemblyProviders"><see cref="IEnumerable{ICanProvideAssemblies}">Providers</see> to provide assemblies</param>
-        /// <param name="assemblyFilters"><see cref="IAssemblyFilters"/> to use for filtering assemblies through</param>
-        /// <param name="assemblyUtility">An <see cref="IAssemblyUtility"/></param>
-        /// <param name="assemblySpecifiers"><see cref="IAssemblySpecifiers"/> used for specifying what assemblies to include or not</param>
-        /// <param name="contractToImplementorsMap"><see cref="IContractToImplementorsMap"/> for keeping track of the relationship between contracts and implementors</param>
+        /// <param name="assemblyProviders">
+        /// <see cref="IEnumerable{ICanProvideAssemblies}">Providers</see> to provide assemblies.
+        /// </param>
+        /// <param name="assemblyFilters">
+        /// <see cref="IAssemblyFilters"/> to use for filtering assemblies through.
+        /// </param>
+        /// <param name="assemblySpecifiers">
+        /// <see cref="IAssemblySpecifiers"/> used for specifying what assemblies to include or not.
+        /// </param>
+        /// <param name="contractToImplementorsMap">
+        /// <see cref="IContractToImplementorsMap"/> for keeping track of the relationship between contracts and implementors.
+        /// </param>
         public AssemblyProvider(
             IEnumerable<ICanProvideAssemblies> assemblyProviders,
             IAssemblyFilters assemblyFilters,
-            IAssemblyUtility assemblyUtility,
             IAssemblySpecifiers assemblySpecifiers,
             IContractToImplementorsMap contractToImplementorsMap)
         {
-            _assemblyProviders = assemblyProviders;
+            var providers = assemblyProviders.ToList();
             _assemblyFilters = assemblyFilters;
-            _assemblyUtility = assemblyUtility;
             _assemblySpecifiers = assemblySpecifiers;
             _contractToImplementorsMap = contractToImplementorsMap;
 
-            HookUpAssemblyAddedForProviders();
-            Populate();
+            foreach (var provider in providers)
+            {
+                provider.AvailableAssemblies.Added += (_, assemblies) => Enqueue(assemblies);
+            }
+
+            Enqueue(providers.SelectMany(p => p.AvailableAssemblies));
         }
 
 #pragma warning disable 1591 // Xml Comments
@@ -73,69 +72,88 @@ namespace Bifrost.Execution
             return _assemblies;
         }
 #pragma warning restore 1591 // Xml Comments
-        void HookUpAssemblyAddedForProviders()
+
+        void Enqueue(IEnumerable<AssemblyInfo> assemblies)
         {
-            foreach (var provider in _assemblyProviders)
+            foreach (var assembly in assemblies)
             {
-                provider.AssemblyAdded += AssemblyLoaded;
+                _assembliesToBeSpecified.Enqueue(assembly);
             }
+
+            Process();
         }
 
-        void AssemblyLoaded(Assembly assembly)
+        // The implementation of this method is complicated by the mechanism that loading an assembly
+        // might trigger an assembly load event of referenced assemblies.
+        void Process()
         {
-            if (assembly.IsDynamic || _assemblies.Contains(assembly))
+            if (Interlocked.CompareExchange(ref _processing, 1, 0) == 1)
             {
                 return;
             }
 
-            var newSpecifiers = _assemblySpecifiers.SpecifyUsingSpecifiersFrom(assembly);
-            AddAssembly(assembly);
-            if (newSpecifiers)
+            try
             {
-                RecalculateAssemblies();
+                DoProcess();
+            }
+            finally
+            {
+                _processing = 0;
             }
         }
 
-        void Populate()
+        void DoProcess()
         {
-            var assemblies = AvailableAssemblies();
-            assemblies.ForEach(a => _assemblySpecifiers.SpecifyUsingSpecifiersFrom(a));
-            assemblies.ForEach(AddAssembly);
-        }
-
-        IList<Assembly> AvailableAssemblies()
-        {
-            return _assemblyProviders
-                .SelectMany(p => p.AvailableAssemblies, (p, ai) => new {p, ai})
-                .GroupBy(d => d.ai.Name)
-                .Select(g => g.First())
-                .Where(d => _assemblyUtility.IsAssembly(d.ai))
-                .Select(d => d.p.Get(d.ai))
-                .Where(a => !a.IsDynamic)
-                .ToList();
-        }
-
-        bool ShouldInclude(Assembly assembly)
-        {
-            return _assemblyFilters.ShouldInclude(new FileInfo(assembly.Location).Name);
-        }
-
-        void AddAssembly(Assembly assembly)
-        {
-            lock (LockObject)
+            do
             {
-                if (!_assemblies.Contains(assembly, _comparer) && ShouldInclude(assembly))
+                var newSpecifications = false;
+                // First, specify from all assemblies in queue.
+                // (Loading an assembly in the queue might add more assemblies to the queue.)
+                while (_assembliesToBeSpecified.Count > 0)
                 {
-                    _assemblies.Add(assembly);
+                    AssemblyInfo next;
+                    if (_assembliesToBeSpecified.TryDequeue(out next) && 
+                        !_specifiedAssemblies.Contains(next, _comparer))
+                    {
+                        var assembly = next.Assembly.Value;
+                        newSpecifications |= _assemblySpecifiers.SpecifyUsingSpecifiersFrom(assembly);
+                        _specifiedAssemblies.Add(next);
+                    }
+                }
+
+                // No more assemblies in queue, add all new assemblies.
+                // (This might again add more assemblies to the queue.)
+                if (newSpecifications)
+                {
+                    AddNewAssemblies();
+                }
+            } while (_assembliesToBeSpecified.Count > 0);
+        }
+
+        void AddNewAssemblies()
+        {
+            foreach (var assembly in _specifiedAssemblies.Select(ai => ai.Assembly.Value))
+            {
+                if (!_assemblies.Contains(assembly) &&
+                    _assemblyFilters.ShouldInclude(new FileInfo(assembly.Location).Name))
+                {
                     _contractToImplementorsMap.Feed(assembly.GetTypes());
+                    _assemblies.Add(assembly);
                 }
             }
         }
 
-        void RecalculateAssemblies()
+        private class AssemblyInfoComparer : IEqualityComparer<AssemblyInfo>
         {
-            var assemblies = AvailableAssemblies();
-            assemblies.ForEach(AddAssembly);
+            public bool Equals(AssemblyInfo x, AssemblyInfo y)
+            {
+                return x.Name == y.Name;
+            }
+
+            public int GetHashCode(AssemblyInfo obj)
+            {
+                return obj.Name.GetHashCode();
+            }
         }
     }
 }